26.06.2025

+++ PrintF-v2/Libft/ft_putnbr_fd.c
@@ -6,13 +6,13 @@
 /*   By: fekandle <fekandle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/10 13:20:11 by fekandle          #+#    #+#             */
-/*   Updated: 2025/06/10 18:51:22 by fekandle         ###   ########.fr       */
+/*   Updated: 2025/06/22 18:12:00 by fekandle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(long n, int fd)
 {
 	if (n == -2147483648)
 		ft_putstr_fd("-2147483648", fd);
